refactor(pdf): isolate preformatted visuals
